chore(tests): move exception listener test
